Beethoven.Allocator.Cruncher (Beethoven v0.3.7)

Crunches signal data and generates an allocation (or busy-score) score for the current node. Signal data is pulled from the Beethoven.Allocator.Ingress.Cache ets table. The aggregated score is stored in the Beethoven.Allocator.Tracker Mnesia table. This PID will refresh the score anytime Beethoven.Allocator.Ingress pushes an update and sends a cast to this PID.

Functions

algorithm(arg)

Crunching algorithm for each signal item.

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

send_check()

@spec send_check() :: :ok

Sends a check message to the local instance of Cruncher

start_link(init_args \\ [])

@spec start_link(any()) :: GenServer.on_start()

Supervisor Entry point.