Beethoven.Allocator (Beethoven v0.3.7)

Allocator is a stack of processes that facilitate the aggregation of telemetry signals to determine how busy a given Beethoven node is.

Public API

allocate/0 Provides the URI of the least-busy node in the cluster.

The busyness of nodes is determined via signals. Refer to the documentation for Beethoven.Allocator.Agent for more information on creating signals.

Functions

allocate()

@spec allocate() :: node()

Provides the URI of the least-busy node in the cluster. Using this function has no side effects so discarding the output without using it will not cause issues.

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

start_link(init_args)

@spec start_link(any()) :: GenServer.on_start()

Supervisor Entry point.